feat: add steam module
