+++ source/client/cl_presence.c
@@ -22,6 +22,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "client.h"
 #include "../gameshared/gs_public.h"
+#include "../steamshim/src/packet_utils.h"
+#include "../steamshim/src/parent/parent.h"
 
 #include "discord_register.h"
 #include "discord_rpc.h"
@@ -512,11 +514,29 @@ static cl_presence_state_t cl_presence_state;
 void UpdatePresenceIfChanged( RichPresence presence )
 {
 	if( memcmp( &cl_presence_state.steam_old_presence, &presence, sizeof( presence ) ) != 0 && Steam_Active()) {
+		char rpc_packet[sizeof( struct buffer_rpc_s ) + 256];
+		struct buffer_rpc_s *req = (struct buffer_rpc_s *)rpc_packet;
+		req->cmd = RPC_SET_RICH_PRESENCE;
 		char steam_display[128];
-		Q_snprintfz(steam_display, sizeof steam_display, "%s | %s", presence.state, presence.details);
-		char* keys[3] = {"score", "details","steam_display"};
-		char* values[3] = {steam_display,presence.details,"#Status_Score"};
-		Steam_SetRichPresence(3, keys, values);
+		Q_snprintfz( steam_display, sizeof steam_display, "%s | %s", presence.state, presence.details );
+
+		{
+			const char *pairs[] = { "score", steam_display };
+			const size_t bufferLen = pack_cstr_null_terminated( (char *)req->buf, 256, pairs, 2 );
+			STEAMSHIM_sendRPC( req, bufferLen + sizeof( struct buffer_rpc_s ), NULL, NULL, NULL);
+		}
+		
+		{
+			const char *pairs[] = { "details", presence.details};
+			const size_t bufferLen = pack_cstr_null_terminated( (char *)req->buf, 256, pairs, 2 );
+			STEAMSHIM_sendRPC( req, bufferLen + sizeof( struct buffer_rpc_s ), NULL, NULL, NULL);
+		}
+		
+		{
+			const char *pairs[] = { "steam_display", "#Status_Score"};
+			const size_t bufferLen = pack_cstr_null_terminated( (char *)req->buf, 256, pairs, 2 );
+			STEAMSHIM_sendRPC( req, bufferLen + sizeof( struct buffer_rpc_s ), NULL, NULL, NULL);
+		}
 		cl_presence_state.steam_old_presence = presence;
 	}
