+++ source/client/client.h
@@ -684,7 +684,6 @@ void CL_Sys_Shutdown( void );
 void Steam_AdvertiseGame( const uint8_t *ip, unsigned short port, uint32_t* syncToken);
 void Steam_GetPersonaName( char *name, size_t namesize );
 const SteamAuthTicket_t* Steam_GetAuthSessionTicketBlocking();
-void Steam_SetRichPresence( int num, const char **key, const char **val );
 void Steam_RequestAvatar(uint64_t steamid, int size);
 void CL_Steam_RunFrame( void );
 void Steam_OpenProfile( uint64_t steamid );
